+++ lib/ecto_shorts/query_builder.ex
@@ -1,15 +1,29 @@
 defmodule EctoShorts.QueryBuilder do
-  @moduledoc "Behaviour for query building from filter tuples"
+  @moduledoc "Behaviour for query building from filter keys and values"
 
-  @type filter_tuple :: {filter_type :: atom, value :: any}
-  @type accumulator_query :: Ecto.Query.t
+  @type filter_key :: atom()
+  @type filter_value :: any()
+  @type query :: Ecto.Query.t()
+  @type queryable :: Ecto.Queryable.t()
 
-  @doc "Adds to accumulator query with filter_type and value"
-  @callback create_schema_filter(filter_tuple, accumulator_query) :: Ecto.Query.t
+  @doc "Adds to the query with the given filter key and value"
+  @callback create_schema_filter(
+    query :: query() | queryable(),
+    filter_key :: filter_key(),
+    filter_value :: filter_value()
+  ) :: query() | queryable()
 
-  @spec create_schema_filter(module, filter_tuple, accumulator_query) :: Ecto.Query.t
-  def create_schema_filter(builder, filter_tuple, query) do
-    builder.create_schema_filter(filter_tuple, query)
+  @doc """
+  Invokes `c:EctoShorts.QueryBuilder.create_schema_filter/3` on the given builder.
+  """
+  @spec create_schema_filter(
+    builder :: module(),
+    query :: query() | queryable(),
+    filter_key :: filter_key(),
+    filter_value :: filter_value()
+  ) :: query() | queryable()
+  def create_schema_filter(builder, query, filter_key, filter_value) do
+    builder.create_schema_filter(query, filter_key, filter_value)
   end
 
   @spec query_schema(Ecto.Queryable.t) :: Ecto.Queryable.t()